+++ Code/Controllers/LYRUIAddressBarViewController.h
@@ -126,13 +126,13 @@
 - (void)selectParticipant:(id<LYRUIParticipant>)participant;
 
 /**
- @abstract Informs the receiver to enter a permanent state. When this method is called, user input and search will be disallowed.
+@abstract Disables user input and searching.
  */
-- (void)setPermanent;
+- (void)disable;
 
 /**
- @abstract A boolean indicating whether or not the reciever is in its permanent state.
+ @abstract A boolean indicating whether or not the receiver is in a disabled state.
  */
-- (BOOL)isPermanent;
+- (BOOL)isDisabled;
 
 @end